move materials & sdf, various xr state and other changes
