+++ packages/editor/src/components/assets/AssetPreviewPanels/ModelPreviewPanel.tsx
@@ -30,10 +30,10 @@ import ResizeObserver from 'resize-observer-polyfill'
 import LoadingView from '@etherealengine/client-core/src/common/components/LoadingView'
 import { setupSceneForPreview } from '@etherealengine/client-core/src/user/components/Panel3D/helperFunctions'
 import { useRender3DPanelSystem } from '@etherealengine/client-core/src/user/components/Panel3D/useRender3DPanelSystem'
-import { SourceType } from '@etherealengine/engine/src/renderer/materials/components/MaterialSource'
-import { removeMaterialSource } from '@etherealengine/engine/src/renderer/materials/functions/MaterialLibraryFunctions'
 import { InfiniteGridHelper } from '@etherealengine/engine/src/scene/classes/InfiniteGridHelper'
 import { ObjectLayers } from '@etherealengine/engine/src/scene/constants/ObjectLayers'
+import { SourceType } from '@etherealengine/engine/src/scene/materials/components/MaterialSource'
+import { removeMaterialSource } from '@etherealengine/engine/src/scene/materials/functions/MaterialLibraryFunctions'
 import { useHookstate } from '@etherealengine/hyperflux'
 
 import { AssetLoader } from '@etherealengine/engine/src/assets/classes/AssetLoader'
