+++ packages/editor/src/components/materials/MaterialLibraryEntry.tsx
@@ -28,11 +28,11 @@ import React, { MouseEvent, StyleHTMLAttributes, useCallback } from 'react'
 import { useDrag } from 'react-dnd'
 
 import { pathResolver } from '@etherealengine/engine/src/assets/functions/pathResolver'
-import { MaterialComponentType } from '@etherealengine/engine/src/renderer/materials/components/MaterialComponent'
-import { MaterialPrototypeComponentType } from '@etherealengine/engine/src/renderer/materials/components/MaterialPrototypeComponent'
-import { MaterialSourceComponentType } from '@etherealengine/engine/src/renderer/materials/components/MaterialSource'
-import { LibraryEntryType } from '@etherealengine/engine/src/renderer/materials/constants/LibraryEntry'
-import { entryId } from '@etherealengine/engine/src/renderer/materials/functions/MaterialLibraryFunctions'
+import { MaterialComponentType } from '@etherealengine/engine/src/scene/materials/components/MaterialComponent'
+import { MaterialPrototypeComponentType } from '@etherealengine/engine/src/scene/materials/components/MaterialPrototypeComponent'
+import { MaterialSourceComponentType } from '@etherealengine/engine/src/scene/materials/components/MaterialSource'
+import { LibraryEntryType } from '@etherealengine/engine/src/scene/materials/constants/LibraryEntry'
+import { entryId } from '@etherealengine/engine/src/scene/materials/functions/MaterialLibraryFunctions'
 import { getMutableState } from '@etherealengine/hyperflux'
 
 import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown'
