+++ packages/engine/src/assets/classes/AssetLoader.ts
@@ -52,8 +52,8 @@ import { EngineState } from '@etherealengine/engine/src/EngineState'
 import { isAbsolutePath } from '../../common/functions/isAbsolutePath'
 import { iOS } from '../../common/functions/isMobile'
 import iterateObject3D from '../../common/functions/iterateObject3D'
-import { SourceType } from '../../renderer/materials/components/MaterialSource'
-import loadVideoTexture from '../../renderer/materials/functions/LoadVideoTexture'
+import { SourceType } from '../../scene/materials/components/MaterialSource'
+import loadVideoTexture from '../../scene/materials/functions/LoadVideoTexture'
 import { DEFAULT_LOD_DISTANCES, LODS_REGEXP } from '../constants/LoaderConstants'
 import { AssetClass } from '../enum/AssetClass'
 import { AssetType } from '../enum/AssetType'
