+++ packages/engine/src/assets/exporters/gltf/extensions/ImageRoutingExtension.ts
@@ -25,7 +25,7 @@ Ethereal Engine. All Rights Reserved.
 
 import { pathJoin, relativePathTo } from '@etherealengine/common/src/utils/miscUtils'
 import { Material, Object3D, Object3DEventMap, Texture } from 'three'
-import { materialFromId } from '../../../../renderer/materials/functions/MaterialLibraryFunctions'
+import { materialFromId } from '../../../../scene/materials/functions/MaterialLibraryFunctions'
 import { pathResolver } from '../../../functions/pathResolver'
 import { GLTFExporterPlugin, GLTFWriter } from '../GLTFExporter'
 import { ExporterExtension } from './ExporterExtension'
