+++ packages/engine/src/assets/loaders/gltf/extensions/RegisterMaterialsExtension.ts
@@ -28,9 +28,9 @@ import { Mesh, Object3D } from 'three'
 import { getState } from '@etherealengine/hyperflux'
 
 import iterateObject3D from '../../../../common/functions/iterateObject3D'
-import { MaterialLibraryState } from '../../../../renderer/materials/MaterialLibrary'
-import { SourceType } from '../../../../renderer/materials/components/MaterialSource'
-import { registerMaterial } from '../../../../renderer/materials/functions/MaterialLibraryFunctions'
+import { MaterialLibraryState } from '../../../../scene/materials/MaterialLibrary'
+import { SourceType } from '../../../../scene/materials/components/MaterialSource'
+import { registerMaterial } from '../../../../scene/materials/functions/MaterialLibraryFunctions'
 import { GLTF, GLTFLoaderPlugin } from '../GLTFLoader'
 import { ImporterExtension } from './ImporterExtension'
