+++ packages/engine/src/renderer/RendererModule.ts
@@ -26,5 +26,6 @@ Ethereal Engine. All Rights Reserved.
 import { DebugRendererSystem } from './DebugRendererSystem'
 import { RenderInfoSystem } from './RenderInfoSystem'
 import { WebGLRendererSystem } from './WebGLRendererSystem'
+import { SDFSystem } from './effects/sdf/SDFSystem'
 
-export { DebugRendererSystem, RenderInfoSystem, WebGLRendererSystem }
+export { DebugRendererSystem, RenderInfoSystem, WebGLRendererSystem, SDFSystem }
