+++ packages/engine/src/renderer/functions/configureEffectComposer.ts
@@ -44,16 +44,16 @@ import { getState } from '@etherealengine/hyperflux'
 
 import { getComponent } from '@etherealengine/ecs/src/ComponentFunctions'
 import { Engine } from '@etherealengine/ecs/src/Engine'
-import { EngineState } from '@etherealengine/engine/src/EngineState'
-import { SDFComponent } from '@etherealengine/engine/src/scene/components/SDFComponent'
 import { ShaderPass } from 'postprocessing'
+import { EngineState } from '../../EngineState'
 import { CameraComponent } from '../../camera/components/CameraComponent'
 import { ObjectLayers } from '../../scene/constants/ObjectLayers'
 import { EffectMap, EffectPropsSchema, Effects } from '../../scene/constants/PostProcessing'
 import { HighlightState } from '../HighlightState'
 import { RendererState } from '../RendererState'
 import { EffectComposerWithSchema, EngineRenderer, PostProcessingSettingsState } from '../WebGLRendererSystem'
-import { SDFShader } from '../effects/SDFShader'
+import { SDFComponent } from '../effects/sdf/SDFComponent'
+import { SDFShader } from '../effects/sdf/SDFShader'
 import { changeRenderMode } from './changeRenderMode'
 
 export const configureEffectComposer = (
