+++ packages/engine/src/renderer/functions/generateNoiseTexture.ts
