+++ packages/engine/src/scene/components/ModelComponent.tsx
@@ -55,8 +55,8 @@ import { AvatarRigComponent } from '../../avatar/components/AvatarAnimationCompo
 import { autoconvertMixamoAvatar, isAvaturn } from '../../avatar/functions/avatarFunctions'
 import { CameraComponent } from '../../camera/components/CameraComponent'
 import { EngineRenderer } from '../../renderer/WebGLRendererSystem'
-import { SourceType } from '../../renderer/materials/components/MaterialSource'
-import { removeMaterialSource } from '../../renderer/materials/functions/MaterialLibraryFunctions'
+import { SourceType } from '../../scene/materials/components/MaterialSource'
+import { removeMaterialSource } from '../../scene/materials/functions/MaterialLibraryFunctions'
 import { addError, removeError } from '../functions/ErrorFunctions'
 import { parseGLTFModel, proxifyParentChildRelationships } from '../functions/loadGLTFModel'
 import { getModelSceneID } from '../functions/loaders/ModelFunctions'
