+++ packages/engine/src/scene/materials/components/MaterialPluginComponent.ts
