+++ packages/engine/src/scene/materials/components/MaterialPrototypeComponent.ts
