+++ packages/engine/src/scene/materials/constants/LibraryEntry.ts
