+++ packages/engine/src/scene/materials/constants/material-prototypes/MeshBasicMaterial.mat.ts
