+++ packages/engine/src/scene/materials/constants/material-prototypes/MeshLambertMaterial.mat.ts
