+++ packages/engine/src/scene/materials/constants/material-prototypes/MeshMatcapMaterial.mat.ts
