+++ packages/engine/src/scene/materials/constants/material-prototypes/MeshToonMaterial.mat.ts
