+++ packages/engine/src/scene/materials/constants/material-prototypes/ShadowMaterial.mat.ts
