+++ packages/engine/src/scene/materials/functions/MaterialFunctions.ts
