+++ packages/engine/src/scene/materials/functions/MaterialLibraryFunctions.ts
