+++ packages/engine/src/scene/materials/functions/MaterialPluginFunctions.ts
