+++ packages/engine/src/scene/materials/functions/bakeToVertices.ts
