+++ packages/engine/src/scene/materials/functions/batchEditMaterials.ts
