+++ packages/engine/src/scene/materials/systems/MaterialLibrarySystem.tsx
