+++ packages/engine/src/xr/XRSessionFunctions.ts
@@ -36,7 +36,6 @@ import { setVisibleComponent } from '../scene/components/VisibleComponent'
 import { TransformComponent } from '../transform/components/TransformComponent'
 import { computeAndUpdateWorldOrigin } from '../transform/updateWorldOrigin'
 import { EngineRenderer } from './../renderer/WebGLRendererSystem'
-import { setTrackingSpace } from './XRScaleAdjustmentFunctions'
 import { ReferenceSpace, XRAction, XRState } from './XRState'
 
 const quat180y = new Quaternion().setFromAxisAngle(new Vector3(0, 1, 0), Math.PI)
@@ -156,7 +155,7 @@ export const getReferenceSpaces = (xrSession: XRSession) => {
     //   if (ReferenceSpace.localFloor && 'addEventListener' in ReferenceSpace.localFloor)
     //     ReferenceSpace.localFloor.addEventListener('reset', onLocalFloorReset, { once: true, passive: true })
     // }
-    setTrackingSpace()
+    XRState.setTrackingSpace()
   }
 
   /** the world origin is an offset to the local floor, so as soon as we have the local floor, define the origin reference space */
